Cleaned up code a bit and working on Search

+++ src/genres/genre.service.ts
@@ -1,8 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Genre } from './genre.model';
 import { GenreRepository } from './genre.repository';
+import { CreateGenreDto } from './dtos/create-genre.dto';
 
 @Injectable()
 export class GenreService {
@@ -19,7 +20,18 @@ export class GenreService {
     return await this.genreRepository.findOne({ where: { id } });
   }
 
-  async create(genre: Genre): Promise<Genre> {
+  async create(createGenreDto: CreateGenreDto): Promise<Genre> {
+    // Check if the genre already exists
+    const existingGenre = await this.genreRepository.findOne({
+      where: { name: createGenreDto.name },
+    });
+  
+    if (existingGenre) {
+      throw new Error(`Genre with name '${createGenreDto.name}' already exists.`);
+    }
+  
+    const genre = new Genre();
+    genre.name = createGenreDto.name;
     return await this.genreRepository.save(genre);
   }
 
@@ -28,7 +40,15 @@ export class GenreService {
     return await this.genreRepository.findOne({ where: { id } });
   }
 
-  async delete(id: number): Promise<void> {
+  async delete(id: number): Promise<{ message: string }> {
+    const genre = await this.genreRepository.findOneBy({ id });
+
+    if(!genre){
+      throw new NotFoundException(`Genre with ID ${id} not found`);
+    }
+
     await this.genreRepository.delete(id);
+
+    return { message: `Genre with ID ${id} has been successfully deleted` }
   }
 }
